chore: remove superchain crate
