+++ crates/genesis/src/superchain/mod.rs
@@ -1,0 +1,16 @@
+//! Contains superchain-specific types.
+
+mod level;
+pub use level::SuperchainLevel;
+
+mod chain;
+pub use chain::Superchain;
+
+mod chains;
+pub use chains::Superchains;
+
+mod config;
+pub use config::SuperchainConfig;
+
+mod info;
+pub use info::SuperchainL1Info;
