+++ crates/registry/Cargo.toml
@@ -16,7 +16,6 @@ workspace = true
 
 [dependencies]
 # Workspace
-maili-superchain.workspace = true
 maili-genesis = { workspace = true, features = ["serde"] }
 
 # Alloy
@@ -31,7 +30,6 @@ lazy_static = { workspace = true, features = ["spin_no_std"] }
 
 [build-dependencies]
 toml = { workspace = true, features = ["parse"] }
-maili-superchain.workspace = true
 serde_json = { workspace = true, features = ["raw_value"] }
 maili-genesis = { workspace = true, features = ["serde"] }
 
@@ -46,6 +44,5 @@ std = [
 	"maili-genesis/std",
 	"serde_json/std",
 	"alloy-primitives/std",
-	"maili-superchain/std",
 	"serde/std"
 ]
