--- crates/registry/superchain/Cargo.toml
@@ -1,36 +1,0 @@
-[package]
-name = "maili-superchain"
-description = "Core superchain types for Maili"
-
-version.workspace = true
-edition.workspace = true
-rust-version.workspace = true
-authors.workspace = true
-license.workspace = true
-homepage.workspace = true
-repository.workspace = true
-exclude.workspace = true
-
-[lints]
-workspace = true
-
-[dependencies]
-# Workspace
-maili-genesis = { workspace = true, features = ["serde"] }
-
-# Alloy
-alloy-primitives = { workspace = true, features = ["map", "map-foldhash"] }
-
-# `serde`
-serde = { workspace = true, features = ["derive", "alloc"] }
-
-[dev-dependencies]
-serde_json.workspace = true
-
-[features]
-default = ["std"]
-std = [
-	"maili-genesis/std",
-	"alloy-primitives/std",
-	"serde/std",
-]
